# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo Camisetas API",
        "description": "API B2B para gestión de camisetas y clientes",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor local Docker"
        }
    ],
    "paths": {
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar todos los clientes",
                "operationId": "de55d50d1b708d9e5c63ee3f8fe8f54a",
                "responses": {
                    "200": {
                        "description": "Lista de clientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre_comercial": {
                                                "type": "string",
                                                "example": "90minutos"
                                            },
                                            "rut": {
                                                "type": "string",
                                                "example": "76.111.111-1"
                                            },
                                            "direccion": {
                                                "type": "string",
                                                "example": "Providencia, Santiago"
                                            },
                                            "categoria": {
                                                "type": "string",
                                                "example": "Preferencial"
                                            },
                                            "contacto_nombre": {
                                                "type": "string",
                                                "example": "Carlos Ruiz"
                                            },
                                            "contacto_email": {
                                                "type": "string",
                                                "example": "carlos@90minutos.cl"
                                            },
                                            "porcentaje_oferta": {
                                                "type": "number",
                                                "example": 15
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear un cliente",
                "operationId": "4ac6cdfa8d6c188fa2cb4f3263d4318a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_comercial",
                                    "rut",
                                    "direccion",
                                    "categoria",
                                    "contacto_nombre",
                                    "contacto_email"
                                ],
                                "properties": {
                                    "nombre_comercial": {
                                        "type": "string",
                                        "example": "NuevaTienda"
                                    },
                                    "rut": {
                                        "type": "string",
                                        "example": "76.333.333-3"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Vitacura, Santiago"
                                    },
                                    "categoria": {
                                        "type": "string",
                                        "example": "Regular",
                                        "enum": [
                                            "Regular",
                                            "Preferencial"
                                        ]
                                    },
                                    "contacto_nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "contacto_email": {
                                        "type": "string",
                                        "example": "juan@nueva.cl"
                                    },
                                    "porcentaje_oferta": {
                                        "type": "number",
                                        "example": 5,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "nombre_comercial": {
                                            "type": "string",
                                            "example": "NuevaTienda"
                                        },
                                        "rut": {
                                            "type": "string",
                                            "example": "76.333.333-3"
                                        },
                                        "categoria": {
                                            "type": "string",
                                            "example": "Regular"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Ver un cliente por ID",
                "operationId": "0d3826d79390929601421b5810b61231",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre_comercial": {
                                            "type": "string",
                                            "example": "90minutos"
                                        },
                                        "rut": {
                                            "type": "string",
                                            "example": "76.111.111-1"
                                        },
                                        "direccion": {
                                            "type": "string",
                                            "example": "Providencia, Santiago"
                                        },
                                        "categoria": {
                                            "type": "string",
                                            "example": "Preferencial"
                                        },
                                        "contacto_nombre": {
                                            "type": "string",
                                            "example": "Carlos Ruiz"
                                        },
                                        "contacto_email": {
                                            "type": "string",
                                            "example": "carlos@90minutos.cl"
                                        },
                                        "porcentaje_oferta": {
                                            "type": "number",
                                            "example": 15
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar un cliente",
                "operationId": "74956b09f3ea120b5dd531c73776ca11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_comercial": {
                                        "type": "string",
                                        "example": "90minutos Plus"
                                    },
                                    "categoria": {
                                        "type": "string",
                                        "example": "Preferencial",
                                        "enum": [
                                            "Regular",
                                            "Preferencial"
                                        ]
                                    },
                                    "porcentaje_oferta": {
                                        "type": "number",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar un cliente",
                "operationId": "d15e304c609752c9f409ab81072fb52e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente eliminado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                }
            }
        },
...
    },
...
}